requirements.txt
beautifulsoup4>=4.12.0,<4.13.0
newspaper3k>=0.2,<0.3
docker>=7.0,<8.0
requests>=2.3,<2.4
requests-file>=2.0, <3.0
docker>=7.0,<8.0
